requirements.txt
streamlit==0.77.0
matplotlib==3.3.4
opencv_python==4.4.0.46
deep_translator==1.10.1
detectron2==0.2.1+cpu
largestinteriorrectangle==0.2.0
manga_ocr==0.1.8
Pillow==9.5.0
